Add Freeze example
