+++ package/src/renderer/components/index.ts
@@ -10,7 +10,6 @@ export * from "./pathEffects";
 export * from "../processors";
 export * from "./Picture";
 
-export * from "./Freeze";
 export * from "./Group";
 export * from "./Mask";
 export * from "./Paint";
